requirements.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-faicon==0.1.0
django-js-asset==1.2.2
django-modeltranslation==0.15.2
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
